+++ src/RegisterSociete/Form/AccountType.php
@@ -5,6 +5,7 @@ namespace App\RegisterSociete\Form;
 use App\Entity\User;
 use App\Form\Custom\RepeatedPasswordType;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
@@ -23,6 +24,11 @@ class AccountType extends AbstractType
             ])
             ->add('nom')
             ->add('password', RepeatedPasswordType::class)
+            ->add('acceptCguCgv', CheckboxType::class, [
+                'label' => 'i_accept_cgu_and_cgv',
+                'label_html' => true,
+                'mapped' => false,
+            ])
         ;
     }