+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Tous les changements apportés au projet sont suivis dans ce fichier.
 Le format est basé sur [Tenez un Changelog](https://keepachangelog.com/fr/1.0.0/).
 
 ## [À venir dans la prochaine version]
+### Ajouté
+- Mails de rappels automatique pour finaliser son inscription
+- Ajout des CGU et CGV
 
 
 
@@ -13,7 +16,6 @@ Le format est basé sur [Tenez un Changelog](https://keepachangelog.com/fr/1.0.0
 ### Ajouté
 - On peut maintenant supprimer un projet
 - Le chef de projet est invité à ajouter des contributeurs dés la création d'un nouveau projet
-- Mails de rappels automatique pour finaliser son inscription
 ### Réparé
 - Empêche un admin de désactiver son propre compte pour ne pas se bloquer